requirements.txt
aioredis==1.3.1
amqp==2.5.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
backcall==0.1.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
decorator==4.4.1
Django==3.0.2
django-appconf==1.0.3
django-compressor==2.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-guardian==2.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
eventlet==0.25.1
googlemaps==4.0.0
greenlet==0.4.15
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
kombu==4.5.0
monotonic==1.5
msgpack==0.6.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.5.2
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
rcssmin==1.0.6
redis==3.2.0
requests==2.22.0
rjsmin==1.1.0
service-identity==18.1.0
six==1.14.0
sorl-thumbnail==12.6.0
sqlparse==0.3.0
traitlets==4.3.3
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.8
zipp==2.1.0
zope.interface==4.7.1
